refactor(invoice): reorganize invoice entity and branding logic for clarity
