+++ src/ui/invoices/create-invoice-form.tsx
@@ -5,8 +5,8 @@ import { FormActionRow } from "@/components/form-action-row";
 import { FormSubmitButton } from "@/components/form-submit-button";
 import { Label } from "@/components/label";
 import type { CustomerField } from "@/features/customers/customer.types";
-import type { InvoiceFormStateCreate } from "@/features/invoices/invoice.actions";
 import { createInvoiceAction } from "@/features/invoices/invoice.actions";
+import type { InvoiceFormStateCreate } from "@/features/invoices/invoice.types";
 import { CustomerSelect } from "@/ui/invoices/customer-select";
 import { InvoiceAmountInput } from "@/ui/invoices/invoice-amount-input";
 import { InvoiceServerMessage } from "@/ui/invoices/invoice-server-message";
